Test doedsdato
