+++ apps/person-search-service/src/main/java/no/nav/registre/testnorge/personsearchservice/adapter/utils/QueryUtils.java
@@ -2,16 +2,10 @@ package no.nav.registre.testnorge.personsearchservice.adapter.utils;
 
 import lombok.experimental.UtilityClass;
 import org.apache.lucene.search.join.ScoreMode;
-import org.elasticsearch.index.query.BoolQueryBuilder;
 import org.elasticsearch.index.query.NestedQueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
-import org.elasticsearch.index.query.RangeQueryBuilder;
 
-import java.time.LocalDate;
 import java.util.Collection;
-import java.util.Optional;
-
-import static java.util.Objects.nonNull;
 
 @UtilityClass
 public class QueryUtils {
@@ -38,31 +32,5 @@ public class QueryUtils {
                 QueryBuilders.existsQuery(path + "." + name),
                 ScoreMode.Avg
         );
-    }
-
-    public static void queryDateField(String path, String field, LocalDate fom, LocalDate tom, BoolQueryBuilder queryBuilder) {
-        getBetween(fom, tom, path + "." + field)
-                .ifPresent(rangeQueryBuilder -> queryBuilder.must(QueryBuilders.nestedQuery(
-                                path,
-                                rangeQueryBuilder,
-                                ScoreMode.Avg
-                        ))
-                );
-    }
-
-    private static Optional<RangeQueryBuilder> getBetween(LocalDate fom, LocalDate tom, String field) {
-        if (fom == null && tom == null) {
-            return Optional.empty();
-        }
-        var builder = QueryBuilders.rangeQuery(field);
-
-        if (nonNull(fom)) {
-            builder.gte(fom);
-        }
-
-        if (nonNull(tom)) {
-            builder.lte(tom);
-        }
-        return Optional.of(builder);
     }
 }
