+++ apps/person-search-service/src/main/java/no/nav/registre/testnorge/personsearchservice/controller/search/Doedsfall.java
@@ -10,7 +10,6 @@ import java.time.LocalDate;
 @AllArgsConstructor
 @NoArgsConstructor(force = true)
 public class Doedsfall {
-    Boolean harIkkeDoedsdato;
     Boolean harDoedsdato;
     LocalDate fom;
     LocalDate tom;
